Print list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,11 +10,11 @@
 
 size_t print_listint(const listint_t *h)
 {
-	size_t elm = 0;
+	int elm = 0;
 
 	while (h)
 	{
-		printf("%u\n", h->n);
+		printf("%d\n", h->n);
 		
 		h = h->next;
 		elm++;
